fix review result

+++ features/transaction/minicart/src/main/java/com/tokopedia/minicart/cartlist/MiniCartListUiModelMapper.kt
@@ -91,7 +91,7 @@ class MiniCartListUiModelMapper @Inject constructor() {
                         bundleId = bundleDetail.bundleID
                         originalPrice = bundleDetail.originalPrice
                         displayPrice = bundleDetail.displayPrice
-                        displayPriceRaw = bundleDetail.displayPriceRaw.toLong()
+                        displayPriceRaw = bundleDetail.displayPriceRaw
                         discountPercentage = bundleDetail.discountPercentage
                         isPreOrder = bundleDetail.isPO
                         isProductsHaveVariant = bundleDetail.isProductsHaveVariant
